backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.11.3",
    "flask-bcrypt>=1.0.1",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-socketio>=5.5.1",
    "openai>=1.99.8",
    "path>=17.1.1",
    "pathlib>=1.0.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "uv>=0.8.4",
]
